Added separate class FileTransport
